requirements.txt
gensim==4.2.0
keybert==0.6.0
matplotlib==3.5.2
numpy==1.21.5
pandas==1.3.5
Pillow==9.2.0
scenedetect==0.6.0.3
scikit_learn==1.1.3
seaborn==0.11.2
spacy==2.3.5
streamlit==1.13.0
torch==1.12.1
torchvision==0.13.1
